refactor(6): extraer cmd_run a jobhunter/pipeline.py
- cmd_run movido as-is con sus 3 fases (scrape, analyze, send)
- Imports explicitos: base64, Rich Progress/Panel/Table/Prompt/Confirm,
  Playwright, agents, scraper, mailer, config, storage, browser, ui
- job.py re-exporta cmd_run
- tests/test_pipeline.py: guard de is_configured=False y sin SESSION_DIR
